move argument parsing to main
