+++ ms2pip/__main__.py
@@ -1,6 +1,8 @@
+import argparse
 import logging
+import multiprocessing
 
-from ms2pip.ms2pipC import (argument_parser, load_configfile, run,
+from ms2pip.ms2pipC import (load_configfile, run,
                             InvalidPEPRECError, NoValidPeptideSequencesError,
                             UnknownOutputFormatError,
                             UnknownFragmentationMethodError,
@@ -22,6 +24,67 @@ ralf.gabriels@ugent.be
 http://compomics.github.io/projects/ms2pip_c.html
     """
     print(logo)
+
+
+def argument_parser():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("pep_file", metavar="<PEPREC file>", help="list of peptides")
+    parser.add_argument(
+        "-c",
+        metavar="CONFIG_FILE",
+        action="store",
+        required=True,
+        dest="config_file",
+        help="config file",
+    )
+    parser.add_argument(
+        "-s",
+        metavar="MGF_FILE",
+        action="store",
+        dest="spec_file",
+        help=".mgf MS2 spectrum file (optional)",
+    )
+    parser.add_argument(
+        "-w",
+        metavar="FEATURE_VECTOR_OUTPUT",
+        action="store",
+        dest="vector_file",
+        help="write feature vectors to FILE.{pkl,h5} (optional)",
+    )
+    parser.add_argument(
+        "-x",
+        action="store_true",
+        default=False,
+        dest="correlations",
+        help="calculate correlations (if MGF is given)",
+    )
+    parser.add_argument(
+        "-p",
+        action="store_true",
+        default=False,
+        dest="match_spectra",
+        help="match peptides to spectra based on predicted spectra (if MGF is given)",
+    )
+    parser.add_argument(
+        "-t",
+        action="store_true",
+        default=False,
+        dest="tableau",
+        help="create Tableau Reader file",
+    )
+    parser.add_argument(
+        "-m",
+        metavar="NUM_CPU",
+        action="store",
+        dest="num_cpu",
+        help="number of CPUs to use (default: all available)",
+    )
+    args = parser.parse_args()
+
+    if not args.num_cpu:
+        args.num_cpu = multiprocessing.cpu_count()
+
+    return args
 
 
 def main():
